Allow multiple a=msid attributes
# Details

As per [RFC 8830](https://datatracker.ietf.org/doc/html/rfc8830), a media section can contain multiple `a=msid` attributes. - This is because the `MediaStreamTrack` being sent in that media section may belong to múltiple `MediaStream`. And hence, the `msid` field must be an array.

Also per spec, `msid` value consists on a mandatory `id` string (which can be an `id` or '-' string) and an optional `appdata`. Within the WebRTC context, `id` is the `id` of the `MediaStream` while `appdata` is the `id` of the `MediaStreamTrack`.

Also adapt the tests.